moved main screen to fragment
