+++ app/build.gradle
@@ -47,6 +47,7 @@ configurations {
 }
 
 dependencies {
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     def nav_version = "2.3.0-beta01"
 
     ktlint 'com.pinterest:ktlint:0.36.0'
