+++ app/src/main/res/navigation/navigation.xml
@@ -1,5 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto" android:id="@+id/navigation">
+    xmlns:app="http://schemas.android.com/apk/res-auto" android:id="@+id/navigation"
+    app:startDestination="@id/taskListFragment">
 
+    <fragment
+        android:id="@+id/taskListFragment"
+        android:name="edu.hm.cs.ma.todoguru.TaskListFragment"
+        android:label="TaskListFragment" >
+        <action
+            android:id="@+id/action_taskListFragment_to_viewTaskFragment"
+            app:destination="@id/viewTaskFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/viewTaskFragment"
+        android:name="edu.hm.cs.ma.todoguru.ViewTaskFragment"
+        android:label="ViewTaskFragment" />
 </navigation>
