+++ app/src/main/res/values/strings.xml
@@ -20,4 +20,6 @@
     <string name="notification_content_description">Have you finished your tasks yet?</string>
     <string name="delete_tasks_short">Delete</string>
     <string name="delete_tasks_long">Delete selected tasks</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
